@@ -5,50 +5,40 @@
   "private": true,
   "proxy": "http://localhost:3001",
   "dependencies": {
-    "@ckeditor/ckeditor5-build-classic": "^29.2.0",
-    "@ckeditor/ckeditor5-react": "^3.0.2",
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
-    "@glidejs/glide": "^3.4.1",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@mui/icons-material": "^5.0.1",
-    "@mui/material": "^5.0.2",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "axios": "^0.21.4",
-    "bootstrap": "^5.1.0",
+    "bootstrap": "^5.2.3",
     "bootstrap-icons": "^1.5.0",
     "cloudinary-react": "^1.7.0",
     "dompurify": "^2.3.3",
     "draft-js": "^0.11.7",
     "draftjs-to-html": "^0.9.1",
     "html-react-parser": "^1.3.0",
+    "immutable": "^4.2.4",
     "jsdom": "^17.0.0",
     "jwt-decode": "^3.1.2",
     "moment": "^2.29.1",
     "newsapi": "^2.4.1",
     "nth-check": "^2.0.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.0.0-beta.6",
     "react-bootstrap-icons": "^1.5.0",
-    "react-dom": "^17.0.2",
-    "react-draft-wysiwyg": "^1.14.7",
-    "react-elastic-carousel": "^0.11.5",
+    "react-dom": "^18.2.0",
+    "react-draft-wysiwyg": "^1.15.0",
     "react-ga": "^3.3.0",
     "react-google-login": "^5.2.2",
-    "react-redux": "^7.2.5",
+    "react-redux": "^8.0.5",
     "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
+    "react-scripts": "^5.0.1",
     "react-share": "^4.4.0",
-    "react-slick": "^0.28.1",
+    "react-slick": "^0.29.0",
     "redux": "^4.1.1",
     "redux-thunk": "^2.3.0",
     "slick-carousel": "^1.8.1",
     "styled-components": "^5.3.1",
     "sweetalert2": "^11.1.9",
-    "sweetalert2-react-content": "^4.2.0",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
