@@ -1,20 +1,21 @@
 import React, { useState } from "react";
 import { useDispatch } from "react-redux";
-import { useHistory } from "react-router-dom";
+import { Link, useHistory } from "react-router-dom";
 import { GoogleLogin } from "react-google-login";
-import {
-  Avatar,
-  Button,
-  Paper,
-  Grid,
-  Typography,
-  Container,
-} from "@material-ui/core";
-import LockOutlinedIcon from "@material-ui/icons/LockOutlined";
-import useStyles from "./styles";
-import Input from "./Input";
-import Icon from "./icon";
-import { signIn, signUp } from "../../actions/auth";
+// import {
+//   Avatar,
+//   Button,
+//   Paper,
+//   Grid,
+//   Typography,
+//   Container,
+// } from "@mui/material";
+// import LockOutlinedIcon from "@mui/icons-material/LockOutlined";
+// import useStyles from "./styles";
+
+import { signUp } from "../../actions/auth";
+import { Google } from "react-bootstrap-icons";
+import { Spinner } from "react-bootstrap";
 const initialStateField = {
   firstName: "",
   lastName: "",
@@ -27,37 +28,41 @@ const Auth = () => {
   const user = JSON.parse(localStorage.getItem("profile"));
   if (user) window.location.replace("/");
 
-  const classes = useStyles();
+  // const classes = useStyles();
   const dispatch = useDispatch();
   const history = useHistory();
   const [formData, setFormData] = useState(initialStateField);
+  const [loading, setLoading] = useState(false);
+  // const [error, setError] = useState(false);
 
-  const [showPassword, setShowPassword] = useState(false);
-  const [showConfirmPassword, setShowConfirmPassword] = useState(false);
-
-  const [isSignup, setIsSignup] = useState(true);
+  // const [isSignup, setIsSignup] = useState(true);
 
   const handleSubmit = (e) => {
     e.preventDefault();
     console.log(formData);
 
-    if (isSignup) {
+    setLoading(true);
+    try {
       dispatch(signUp(formData, history));
-    } else {
-      dispatch(signIn(formData, history));
+    } catch (e) {
+      console.log(e);
+
+      setLoading(false);
     }
+
+    setLoading(false);
+
+    // if (isSignup) {
+    //   dispatch(signUp(formData, history));
+    // } else {
+    //   dispatch(signIn(formData, history));
+    // }
   };
   const handleChange = (e) => {
     // e.preventDefault()
     setFormData({ ...formData, [e.target.name]: e.target.value });
   };
-  const switchSignUp = () => setIsSignup((prevIsSignup) => !prevIsSignup);
-
-  const handleShowPassword = () =>
-    setShowPassword((prevShowPassword) => !prevShowPassword);
-
-  const handleShowConfirmPassword = () =>
-    setShowConfirmPassword((prevShowPassword) => !prevShowPassword);
+  // const switchSignUp = () => setIsSignup((prevIsSignup) => !prevIsSignup);
 
   const googleSuccess = async (res) => {
     const result = res?.profileObj;
@@ -77,94 +82,77 @@ const Auth = () => {
 
   return (
     <>
-      <Container component="main" maxWidth="xs">
+      <main className="container d-flex flex-column align-items-center justify-content-center">
         <br />
         <h2>WELCOME TO GOODNEWS!</h2>
-        <Paper className={classes.paper} elevation={3}>
-          <Avatar className={classes.avatar}>
-            <LockOutlinedIcon />
-          </Avatar>
-          <Typography variant="h5">
-            {isSignup ? "Sign Up" : "Sign In"}
-          </Typography>
-          <form className={classes.form} onSubmit={handleSubmit}>
-            <Grid container spacing={2}>
-              {isSignup && (
-                <>
-                  <Input
-                    name="firstName"
-                    label="First Name"
-                    handleChange={handleChange}
-                    autoFocus
-                    half
-                  />
-                  <Input
-                    name="lastName"
-                    label="Last Name"
-                    handleChange={handleChange}
-                    half
-                  />
-                </>
-              )}
-              <Input
+        <main className="col-8">
+          <h3> Sign In</h3>
+          <form className="d-flex flex-column" onSubmit={handleSubmit}>
+            <section>
+              <input
+                name="firstName"
+                placeholder="First Name"
+                onChange={handleChange}
+                autoFocus
+                className="form-control mt-3"
+              />
+              <input
+                name="lastName"
+                placeholder="Last Name"
+                className="form-control mt-3"
+                onChange={handleChange}
+              />
+
+              <input
                 name="email"
-                label="Email Address"
-                handleChange={handleChange}
                 type="email"
+                placeholder="Email Address"
+                onChange={handleChange}
+                className="form-control mt-3"
               />
-              <Input
+              <input
                 name="password"
-                label="Password"
-                handleChange={handleChange}
-                type={showPassword ? "text" : "password"}
-                handleShowPassword={handleShowPassword}
+                placeholder="Password"
+                className="form-control mt-3"
+                onChange={handleChange}
+                // type={showPassword ? "text" : "password"}
+                // handleShowPassword={handleShowPassword}
               />
-              {isSignup && (
-                <Input
-                  name="confirmPassword"
-                  label="Repeat Password"
-                  handleChange={handleChange}
-                  type={showConfirmPassword ? "text" : "password"}
-                  handleShowPassword={handleShowConfirmPassword}
-                />
-              )}
-            </Grid>
-            <Button
-              type="submit"
-              fullWidth
-              variant="contained"
-              color="primary"
-              className={classes.submit}
-            >
-              {isSignup ? "Sign Up" : "Sign In"}
-            </Button>
+
+              <input
+                name="confirmPassword"
+                placeholder="Repeat Password"
+                className="form-control mt-3"
+                onChange={handleChange}
+                // type={showConfirmPassword ? "text" : "password"}
+                // handleShowPassword={handleShowConfirmPassword}
+              />
+            </section>
+            <button type="submit" className="main-btn mt-5">
+              {loading && <Spinner />} <h2> Sign Up</h2>
+            </button>
             <GoogleLogin
               clientId="605636590964-tj8lo0qpu02vfkofbuv79d247j0rb9bs.apps.googleusercontent.com"
               render={(renderProps) => (
-                <Button
-                  className={classes.googleButton}
-                  color="primary"
-                  fullWidth
+                <button
+                  className="mt-3"
                   onClick={renderProps.onClick}
                   disabled={renderProps.disabled}
-                  startIcon={<Icon />}
-                  variant="contained"
                 >
-                  Google Sign In
-                </Button>
+                  Google Sign In <Google size={30} color="var(--main-color)" />
+                </button>
               )}
               onSuccess={googleSuccess}
               onFailure={googleFailure}
               cookiePolicy="single_host_origin"
             />
-            <Button onClick={switchSignUp}>
-              {isSignup
-                ? "Already Have an account? Sign In"
-                : "Don't have an account? click to Sign Up!"}
-            </Button>
+
+            <p className="mt-2">
+              Already Have an account? <Link to="/authsignin">Sign In</Link>
+            </p>
           </form>
-        </Paper>
-      </Container>
+        </main>
+      </main>
     </>
   );
 };
