@@ -1,33 +1,42 @@
-import React from 'react'
-import { TextField, Grid, InputAdornment, IconButton } from '@material-ui/core'
-import Visibility from '@material-ui/icons/Visibility'
-import VisibilityOff from '@material-ui/icons/VisibilityOff'
+import React from "react";
+// import { TextField, Grid, InputAdornment, IconButton } from "@mui/material";
+// import Visibility from "@mui/icons-material/Visibility";
+// import VisibilityOff from "@mui/icons-material/VisibilityOff";
 
+const Input = ({
+  name,
+  handleChange,
+  label,
+  autoFocus,
+  type,
+  handleShowPassword,
+  half,
+}) => {
+  return (
+    <main item xs={12} sm={half ? 6 : 12}>
+      <textarea
+        name={name}
+        onChange={handleChange}
+        variant="outlined"
+        required
+        fullWidth
+        label={label}
+        autoFocus={autoFocus}
+        type={type}
+        InputProps={
+          name === "password" || name === "confirmPassword"
+            ? {
+                endAdornment: (
+                  <button onClick={handleShowPassword}>
+                    {type === "password" ? <div> i </div> : <div> e</div>}
+                  </button>
+                ),
+              }
+            : null
+        }
+      />
+    </main>
+  );
+};
 
-const Input = ({name, handleChange, label, autoFocus, type, handleShowPassword, half}) => {
-    return (
-        <Grid item xs={12} sm={ half ? 6 : 12 }>
-            <TextField 
-                name={name}
-                onChange={handleChange}
-                variant="outlined"
-                required
-                fullWidth
-                label={label}
-                autoFocus={autoFocus}
-                type={type}
-                InputProps={ (name === 'password' || name === 'confirmPassword') ? {
-                    endAdornment: (
-                        <InputAdornment position="end">
-                            <IconButton onClick={handleShowPassword}>
-                                {type === 'password' ? <Visibility /> : <VisibilityOff />}
-                            </IconButton>
-                        </InputAdornment>
-                    ),
-                } : null}
-            />
-        </Grid>
-    )
-}
-
-export default Input
+export default Input;
