@@ -1,40 +1,40 @@
-import { makeStyles } from '@material-ui/core/styles';
-import { lightGreen } from '@material-ui/core/colors';
+import { makeStyles } from "@mui/styles";
+import { lightGreen } from "@mui/material/colors";
 
 export default makeStyles((theme) => ({
   appBar: {
     borderRadius: 15,
-    margin: '30px 0',
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    padding: '10px 50px',
+    margin: "30px 0",
+    display: "flex",
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "center",
+    padding: "10px 50px",
   },
   heading: {
-    color: 'rgba(0,183,255, 1)',
-    textDecoration: 'none',
+    color: "rgba(0,183,255, 1)",
+    textDecoration: "none",
   },
   image: {
-    marginLeft: '15px',
+    marginLeft: "15px",
   },
   toolbar: {
-    display: 'flex',
-    justifyContent: 'flex-end',
-    width: '400px',
+    display: "flex",
+    justifyContent: "flex-end",
+    width: "400px",
   },
   profile: {
-    display: 'flex',
-    justifyContent: 'space-between',
-    width: '400px',
+    display: "flex",
+    justifyContent: "space-between",
+    width: "400px",
   },
   userName: {
-    display: 'flex',
-    alignItems: 'center',
+    display: "flex",
+    alignItems: "center",
   },
   brandContainer: {
-    display: 'flex',
-    alignItems: 'center',
+    display: "flex",
+    alignItems: "center",
   },
   green: {
     color: theme.palette.getContrastText(lightGreen[500]),
