@@ -2,6 +2,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 import NewsCard from "./NewsCard";
 import Styled from "styled-components";
+import { Spinner } from "react-bootstrap";
 
 const Heading = Styled.div`
 
@@ -17,9 +18,8 @@ background-color: var(--main-color);
 
 `;
 
-function HomepagePost(props) {
-  const { data } = props;
-  console.log(data.posts);
+function HomepagePost({ data }) {
+  // console.log(data.posts);
 
   return (
     <div className="d-flex flex-column justify-content-between">
@@ -35,23 +35,27 @@ function HomepagePost(props) {
       <hr />
 
       <section className="d-flex flex-wrap">
-        {data.posts.slice(0, 3).map((news) => (
-          <div className="col-12 col-md-6 col-lg-4 ">
-            <NewsCard
-              key={news._id}
-              title={news.title}
-              name={news.nameOfAuthor}
-              imagesrc={news.imageUrl}
-              description={news.description.slice(0, 150)}
-              url={"/post?id=" + news._id}
-              likes={news.numberOfLikes}
-              views={news.numberOfViews}
-              comment={news.comments.length}
-              postId={news._id}
-              likeArray={news.likes}
-            />
-          </div>
-        ))}
+        {!data ? (
+          <Spinner />
+        ) : (
+          data.posts.slice(0, 3).map((news) => (
+            <div key={news._id} className="col-12 col-md-6 col-lg-4 ">
+              <NewsCard
+                key={news._id}
+                title={news.title}
+                name={news.nameOfAuthor}
+                imagesrc={news.imageUrl}
+                description={news.description.slice(0, 150)}
+                url={"/post?id=" + news._id}
+                likes={news.numberOfLikes}
+                views={news.numberOfViews}
+                comment={news.comments.length}
+                postId={news._id}
+                likeArray={news.likes}
+              />
+            </div>
+          ))
+        )}
       </section>
     </div>
   );
