@@ -108,7 +108,7 @@ export const MenuData = [
 
 //   React.useEffect(() => {
 //     axios
-//       .get("https://api-good-news.herokuapp.com/api/categories/all")
+//       .get("https://goodnews-ng.herokuapp.com/api/categories/all")
 //       .then((response) => {
 //         setCategories(response.data);
 //       });
