@@ -1,29 +1,30 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import React from "react";
+import ReactDOM from "react-dom/client";
+import "./index.css";
+import App from "./App";
 
-import { Provider } from 'react-redux'
-import { createStore, applyMiddleware, compose } from 'redux'
+import { Provider } from "react-redux";
+import {
+  legacy_createStore as createStore,
+  applyMiddleware,
+  compose,
+} from "redux";
 
-import thunk from 'redux-thunk'
-import reducers from './reducers'
+import thunk from "redux-thunk";
+import reducers from "./reducers";
 
-const store = createStore(reducers, compose(applyMiddleware(thunk)))
+const store = createStore(reducers, compose(applyMiddleware(thunk)));
 
+const root = ReactDOM.createRoot(document.getElementById("root"));
 
-
-ReactDOM.render(
+root.render(
   <Provider store={store}>
-  <React.StrictMode>
+    <React.StrictMode>
       <App />
-  </React.StrictMode>
-  </Provider>,
-  document.getElementById('root')
+    </React.StrictMode>
+  </Provider>
 );
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
