@@ -7,7 +7,7 @@ import Newscard from "../components/NewsCard";
 import Pagination from "../components/Pagination";
 // bootsrapt css
 import "bootstrap/dist/css/bootstrap.min.css";
-import { CircularProgress } from "@material-ui/core";
+import { Spinner } from "react-bootstrap";
 
 const Pagins = Styled.div`
 
@@ -18,7 +18,7 @@ justify-content: center;
 `;
 
 // import axios from "axios";
-const baseURL = "https://api-good-news.herokuapp.com/api";
+const baseURL = "https://goodnews-ng.herokuapp.com/api";
 
 function getQuery() {
   let search = window.location.search;
@@ -71,7 +71,7 @@ function Categories() {
   return (
     <>
       {!categories.length ? (
-        <CircularProgress />
+        <Spinner />
       ) : (
         <div>
           <section className="container d-flex flex-wrap">
